apps/python/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel", "cmake>=3.12"]
build-backend = "setuptools.build_meta"

[project]
name = "connext-dds-python-app"
version = "1.0.0"
description = "RTI Connext DDS Python application with automated code generation"
readme = "README.md"
license = {text = "RTI License"}
authors = [
    {name = "Real-Time Innovations", email = "support@rti.com"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.6",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
requires-python = ">=3.6,<3.13"
dependencies = [
    "rti.connext==7.3.0",
    # RTI Connext Python API - REQUIRED for running this application
    # Installation guide: https://community.rti.com/static/documentation/connext-dds/current/doc/manuals/connext_dds_professional/installation_guide/installing.html#installing-python-c-or-ada-packages
    # 
    # Note: This will be automatically installed when you run 'pip install -e .'
    # The included install.sh script also handles RTI Python API installation
]

# No optional dependencies currently defined

# [project.scripts]
# No global scripts - run example_io_app directly

[project.urls]
Homepage = "https://www.rti.com"
Repository = "https://github.com/rticommunity/rticonnextdds-examples"
Documentation = "https://community.rti.com/static/documentation/connext-dds/7.3.0/"

[tool.setuptools]
packages = ["codegen"]
package-dir = {"codegen" = "../../dds/python/codegen"}

[tool.setuptools.package-data]
codegen = ["*.py"]

# Custom build hook to run CMake codegen before package build
[tool.setuptools.cmdclass]
build_py = "build_hooks.BuildPyWithCodegen"

# No additional tool configurations needed